Refactor TensorRT utilization tools
