+++ yolort/relaying/logits_decoder.py
@@ -1,4 +1,5 @@
 # Copyright (c) 2021, yolort team. All rights reserved.
+
 from typing import List, Tuple
 
 import torch
